requirements.txt
biopython==1.85
h5py==3.13.0
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5
rdkit==2025.3.3
fair-esm==2.0.0
mordred==1.2.0
optuna==4.3.0
scikit-learn==1.3.2
tqdm==4.67.1
torch==2.6.0
requests==2.32.5
huggingface-hub==0.33.1
lightgbm==4.6.0
joblib==1.5.2
matplotlib==3.10.6
scipy==1.15.3
sentencepiece==0.2.0
setuptools==80.9.0
tokenizers==0.21.2
transformers==4.53.0
timeout_decorator==0.5.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127